chore: extract bash readonly approval functions to smaller module
